- restructure, keep cryptography as one package
