+++ src/main/kotlin/penguin/penground/cryptography/controller/CryptographyApiController.kt
@@ -1,9 +1,9 @@
-package penguin.penground.controller
+package penguin.penground.cryptography.controller
 
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.RestController
 import penguin.penground.api.CryptographyApi
-import penguin.penground.api.CryptographyApiHandler
+import penguin.penground.cryptography.api.CryptographyApiHandler
 import penguin.penground.model.DecryptionRequest
 import penguin.penground.model.DecryptionResponse
 import penguin.penground.model.EncryptionRequest
