+++ src/main/kotlin/penguin/penground/cryptography/service/impl/LkhCryptographyAlgorithmService.kt
@@ -1,13 +1,13 @@
-package penguin.penground.service.cryptography.impl
+package penguin.penground.cryptography.service.impl
 
 import org.springframework.stereotype.Service
 import penguin.penground.model.CryptographyAlgorithm
-import penguin.penground.service.cryptography.CryptographyAlgorithmAction
+import penguin.penground.cryptography.service.CryptographyAlgorithmService
 import java.util.*
 import kotlin.random.Random
 
 @Service
-class LkhCryptographyAlgorithmAction : CryptographyAlgorithmAction {
+class LkhCryptographyAlgorithmService : CryptographyAlgorithmService {
     companion object {
         private const val END_OF_PAYLOAD = "JKC01**"
         private const val MAX_NUM_OF_ITERATION = 20
